+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Removed metata elements which are immutable from the module parameters list;
+- Removed metadata elements which are immutable from the module parameters list;
   they will still be included in the return value if set on a zone.
+
+- pdns_auth_tsigkey.py module for managing TSIG keys.
 
 ## 1.0.0 - 2020-04-03